Clean: imports
Consolidated the monitor.rs with uniswap.rs
