+++ crates/clairvoyance/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod tokens;
 pub mod uniswap;
 pub mod utils;
-pub mod monitor;
