+++ crates/cli/src/main.rs
@@ -1,7 +1,7 @@
 use clairvoyance::tokens::get_tokens;
-use clairvoyance::uniswap::Pool;
+use clairvoyance::uniswap::{Pool, get_pool};
 use clairvoyance::utils::get_provider;
-use clairvoyance::monitor::get_pool;
+
 
 use clap::{Parser, Subcommand};
 use ethers::prelude::*;
@@ -69,20 +69,13 @@ async fn main() -> Result<()> {
                 Some(config) => {
                     let config_obj = config::Config::new(config.clone());
                     println!("{:#?}", config);
-                    // get_pool(token0: &str, token1: &str, bp: String) -> Result<Pool, Box<dyn Error>>
-                    // pool = get_pool(token0, token1, bp).await.unwrap()
-                    // get_pool(token0, token1, bp).await.unwrap()
+                    // get pool with stuff from config
+                    // pool = get_pool(token0, token1, bp);
                 }
                 None => {
-                    // let tokens = get_tokens();
-                    // let token0 = tokens.get(token0).unwrap();
-                    // let token1 = tokens.get(token1).unwrap();
-                    // let bp = bp.parse::<u32>().unwrap();
-        
-        
-                    let pool = get_pool(token0, token1, bp).unwrap();
-                    
-        
+                    // get pool with stuff from CLI/Defaults
+                    pool = get_pool(token0, token1, bp).unwrap();
+            
                     let pools = [pool];
         
                     for pool in pools {
@@ -94,5 +87,7 @@ async fn main() -> Result<()> {
         }
         None => {}
     }
+
     Ok(())
+
 }
